imagen/requirements.txt
Flask
Flask-RESTful
pillow
